renamed app_test.js > app.js
