+++ app.js
@@ -1,305 +1,320 @@
-"use strict";
-var DigitalOceanApi = require('doapi');
-var fs = require('fs');
+"use strict"
+var DigitalOceanApi = require('doapi')
+var fs = require('fs')
 var node_ssh = require('node-ssh')
-var inquirer = require('inquirer');
+var inquirer = require('inquirer')
 
 var readable = require('stream').Readable
 
-var client;
-var sshKey;
+var client
+var sshKey
 var ssh = new node_ssh()
 
-var configurations = {};
-var prompt = inquirer.createPromptModule();
-var promptQuestions = [
-    {
-        type: 'input',
-        name: 'do_api_token',
-        message: 'Digital Ocean token:'
-    }, {
-        type: 'input',
-        name: 'ssh_key_location',
-        message: 'Name of SSH key in ~/.ssh/ to connect to Streisand droplet:'
-    }
-]
-var sequestOpts = {
-    privateKey: sshKey,
-    readyTimeout: 20000
+var configurations = {
+  shadowsocks: {
+    'config.json': '/etc/shadowsocks-libev/config.json'
+  }
 };
-var time = new Date();
-var targetRecord;
+const prompt = inquirer.createPromptModule()
+const promptQuestions = [{
+  type: 'input',
+  name: 'do_api_token',
+  message: 'Digital Ocean token:'
+}, {
+  type: 'input',
+  name: 'ssh_key_location',
+  message: 'Name of SSH key in ~/.ssh/ to connect to Streisand droplet:'
+}, {
+  type: 'input',
+  name: 'domain',
+  message: 'domain tied to Streisand:'
+}]
+var time = new Date()
 
 //gather required info from user
 function getSessionInfo() {
-    var sessionInfo = {
-        do_api_token: null,
-        ssh_key_location: null
-    }
+  var sessionInfo = {
+    do_api_token: null,
+    ssh_key_location: null
+  }
 
-    return prompt(promptQuestions).then(answers => {
-        sessionInfo.do_api_token = answers.do_api_token
-        sessionInfo.ssh_key_location = answers.ssh_key_location
-        sshKey = process.env.HOME + '/.ssh/' + sessionInfo.ssh_key_location
-        client = new DigitalOceanApi({token: sessionInfo.do_api_token});
-        console.log(sessionInfo)
-        return sessionInfo;
-    }).catch(error => console.log(error))
+  return prompt(promptQuestions).then(answers => {
+    sessionInfo.do_api_token = answers.do_api_token
+    sessionInfo.ssh_key_location = answers.ssh_key_location
+    sessionInfo.domain = answers.domain
+    sshKey = process.env.HOME + '/.ssh/' + sessionInfo.ssh_key_location
+    client = new DigitalOceanApi({
+      token: sessionInfo.do_api_token
+    })
+    return sessionInfo
+  }).catch(error => console.log(error))
 }
 
 //find the latest instance of streisand from Digital Ocean and store it.
 function getCurrentServer(sessionInfo) {
-    return client.dropletGetAll().then(droplets => {
-        return droplets.filter(el => {
-            if (el.name.search('streisand-') !== -1) {
-                return el.name; 
-            }
-        });
-    }).then(droplet => {
-        sessionInfo.old_server = droplet[0];
-        console.log('The current server is: ' + sessionInfo.old_server.name +
-          " " + sessionInfo.old_server.networks.v4[0].ip_address)
-        return sessionInfo
-    }).catch(error => console.error(error))
+  console.log('finding current instance of Streisand. . .')
+  return client.dropletGetAll().then(droplets => {
+    return droplets.filter(el => {
+      if (el.name.search('streisand-') !== -1) {
+        return el.name
+      }
+    })
+  }).then(droplet => {
+    sessionInfo.old_server = droplet[0];
+    console.log('The current server is: ' + sessionInfo.old_server.name +
+      " " + sessionInfo.old_server.networks.v4[0].ip_address)
+    return sessionInfo
+  }).catch(error => console.error(error))
 }
 
 //using the return from getCurrentServer, start a new droplet using the snapshot of the old server.
 function startNewServer(sessionInfo) {
-    var newServerDetails = {
-        "name": "streisand-" + (time.getMonth() + 1) + "-" + time.getDate() +
-          "-" + time.getHours() + "." + time.getMinutes(),
-        "region": "sfo2",
-        //"region": oldServer.oldserver.region.slug,
-        "size": "512mb",
-        "image": sessionInfo.old_server.snapshot_ids[0],
-        "ssh_keys": [1278744]
-    };
+  var image
+  if(!sessionInfo.old_server.snapshot_ids[0]){
+    console.log('no snapshot present, using image')
+    image = sessionInfo.old_server.image.id
+  } else image = sessionInfo.old_server.snapshot_ids[0]
 
-    return client.dropletNew(newServerDetails).then(newDroplet => {
-        newServerDetails = newDroplet;
-        var numAttempts = 0;
+  var newServerDetails = {
+    "name": "streisand-" + (time.getMonth() + 1) + "-" + time.getDate() +
+      "-" + time.getHours() + "." + time.getMinutes(),
+    "region": "sfo2",
+    //"region": oldServer.oldserver.region.slug,
+    "size": "512mb",
+    "image": image,
+    "ssh_keys": [1278744]
+  }
+  console.log('starting new droplet. . .')
 
-        function getIP() {
+  return client.dropletNew(newServerDetails).then(newDroplet => {
+    newServerDetails = newDroplet
+    var numAttempts = 0
 
-            return client.dropletGet(newServerDetails.id).then(startingDroplet => {
-                numAttempts++
-                console.log("Find IP attempt # ", numAttempts);
+    function getIP() {
 
-                if (startingDroplet.networks.v4[0].ip_address != undefined) {
-                    newServerDetails = startingDroplet;
-                    console.log("newServer updated with IPv4 address: ",
-                      startingDroplet.networks.v4[0].ip_address);
-                    return newServerDetails;
+      return client.dropletGet(newServerDetails.id).then(startingDroplet => {
+        numAttempts++
+        console.log("Find IP attempt # ", numAttempts)
 
-                } else {
-                    if (numAttempts > 15) {
-                        console.log("failed to reach server after " + numAttempts
-                          + " attempts");
-                    } else {
-                        return getIP();
-                    }
-                }
-            }).then(server => {
-                sessionInfo.new_server = server
-                console.log(sessionInfo.new_server.name + " " +
-                  sessionInfo.new_server.networks.v4[0].ip_address)
-                return sessionInfo
-            })
+        if (startingDroplet.networks.v4[0].ip_address != undefined) {
+          newServerDetails = startingDroplet;
+          console.log("newServer updated with IPv4 address: ",
+            startingDroplet.networks.v4[0].ip_address);
+          return newServerDetails;
+
+        } else {
+          if (numAttempts > 15) {
+            console.log("failed to reach server after " + numAttempts +
+              " attempts");
+          } else {
+            return getIP()
+          }
         }
-        return getIP();
-    }).catch(error => console.log(error))
+      }).then(server => {
+        sessionInfo.new_server = server
+        console.log(sessionInfo.new_server.name + " " +
+          sessionInfo.new_server.networks.v4[0].ip_address)
+        return sessionInfo
+      })
+    }
+    return getIP()
+  }).catch(error => console.log(error))
 }
 
-function checkConnection(sessionInfo){
-  var attempts = 25
-  console.log('Attempting to connect to server ' + sessionInfo.new_server.name + '. . . ');
-  console.log('Attempt: #1');
-  function connect(attempt) {
-    ssh.connect({
+function checkConnection(sessionInfo) {
+  const execUptime = function (sessionInfo) {
+    return ssh.connect({
       host: sessionInfo.new_server.networks.v4[0].ip_address,
       username: 'root',
       privateKey: sshKey
-    }).then( results => {
-      return ssh.execCommand('uptime')
-    }, rejection => {
-      attempts--
-      console.error('failed, ' + attempts +  ' attempts left . . .');
-      if (attempts > 0) {
-      connect(attempts);
-    }
-    }).then(results => {
-      console.log('STDOUT: ' + results.stdout)
-      console.log('STDERR: ' + results.stderr)
-      return console.log('successfully connected');
-    }).catch(error => console.error(error))
+    }).then(() => ssh.execCommand('uptime'))
   }
-  connect(attempts)
+
+  function tryConnect() {
+    let attempt = 0
+
+    function loop() {
+      console.log('Attempting to connect to server ' + sessionInfo.new_server.name + '. . . ')
+      console.log('Attempt: ' + (attempt + 1))
+      return execUptime(sessionInfo).then(results => {
+        if (results.hasOwnProperty('stdout') || results.hasOwnProperty('stderr')) {
+          if (results.hasOwnProperty('stdout')) {
+            console.log('STDOUT: ' + results.stdout)
+            return sessionInfo
+          } else if (results.hasOwnProperty('stderr')) {
+            console.log('command failed - STDERR: ' + results.stderr)
+            attempt++
+            if (attempt !== 25) {
+              console.log('trying again . . . ');
+              return loop()
+            }
+          }
+        } else {
+          console.log('unknown failure: ' + results)
+        }
+      }).catch(error => {
+        console.log('failed to connect . . . ');
+        attempt++
+        console.error(error)
+        if (attempt !== 25) {
+          console.log('trying again . . . ');
+          return loop()
+        }
+        console.log('failed to connect after 25 attempts')
+      })
+    }
+    return loop()
+  }
+  return tryConnect()
 }
 
+function getShadowsocks(sessionInfo) {
+  console.log('updating Shadowsocks. . . ')
 
-
-var test = ssh.connect({
-  host: '138.197.193.189',
-  username: 'root',
-  privateKey: process.env.USERPROFILE + '\\repos\\juke\\do'
-}).then( results => {
-  return ssh.execCommand('cat', ['/etc/shadowsocks-libev/config.json'])
-}).then(result => {
-  return console.log('STOUT' + result)
-}).catch(error => console.log(error))
-
-test 
-//getSessionInfo().then(getCurrentServer).then(startNewServer).then(checkConnection)
-
-//CODE YET TO BE CONVERTED AND IMPLEMENTED
-/*
-//try to ssh and run `uptime` in a loop to determine when the droplet is available
-function checkConnection(serverIP, attempts) {
-  sequestOpts.command = 'uptime';
-  var results = function (error, stdout) {
-    if (error) {
-      console.error(error);
-      checkConnection(serverIP, attempts - 1);
-    } else {
-      console.log('sucessfully connected to server' + stdout);
+  function readConfig(service, config) {
+    sessionInfo[service] = {}
+    sessionInfo[service][config] = config
+    const streamOpts = {
+      flags: 'r',
+      encoding: 'utf-8',
+      handle: null,
+      mode: 0o666,
+      autoClose: true
     }
-  }
 
-  if (attempts < 1) {
-    console.log('Unable to connect to ' + serverIP + 'Aborting.');
-  } else {
-    console.log('attempting to connect to ' + serverIP);
-    sequest('root@' + serverIP, sequestOpts, results);
-  }
-};
+    var streamData = []
 
-//copy Shadowsocks config from old server onto new server
-function getShadowsocks() {
-  var oldConfigSSH = sequest.get('root@' + oldServer[0].networks.v4[0].ip_address,
-  '/etc/shadowsocks-libev/config.json', sequestOpts, sequestResponse);
-  configurations.shadowsocks = {};
-
-  function updateShadowsocks() {
-    console.log("updateShadowsocks callback started.");
-    configurations.shadowsocks.server = newServer.networks.v4[0].ip_address;
-    //this is a real hack that requires success, rework.
-    updateConfig(serviceRestart);
-  };
-
-  function updateConfig(callback) {
-    var configReadable = new readable;
-    var writer = sequest.put('root@' + newServer.networks.v4[0].ip_address,
-    '/etc/shadowsocks-libev/config.json', {
-      mode: '0640',
-      privateKey: sshKey,
-      readyTimeout: 20000
-    });
-
-    configReadable.push(JSON.stringify(configurations.shadowsocks, null, 2) + '\n');
-    configReadable.push(null);
-    configReadable.pipe(writer);
-    writer.on('close', function () {
-      console.log("finished writing");
-      callback('shadowsocks-libev');
+    return ssh.connect({
+      host: sessionInfo.old_server.networks.v4[0].ip_address,
+      username: 'root',
+      privateKey: sshKey
+    }).then(session => {
+      return session.requestSFTP().then(sftp => {
+        const readStream = sftp.createReadStream(configurations[service][config], streamOpts)
+        const stream = new Promise((resolve, reject) => {
+          readStream.on('data', data => {
+            streamData.push(data)
+          })
+          readStream.on('end', data => {
+            console.log(config + ' was successfully downloaded . . . ')
+            sessionInfo[service][config] = JSON.parse(streamData)
+            sessionInfo[service][config].server = sessionInfo.new_server.networks.v4[0].ip_address
+            resolve()
+          })
+          readStream.on('error', () => reject())
+        })
+        return stream
+      })
+    }).catch(error => {
+      return console.error(error);
     })
-  };
-  streamToString(oldConfigSSH, configurations, 'shadowsocks', updateShadowsocks);
-};
+  }
 
-function destroyOldServer(dropletID) {
-  return client.dropletDestroy(dropletID).then(function (destroyedDroplet) {
-    return console.log(destroyedDroplet);
-  });
-};
+  function writeConfig(service, config) {
+    const streamOpts = {
+      flags: 'w',
+      encoding: 'utf-8',
+      handle: null,
+      mode: 0o666,
+      autoClose: true
+    }
+
+    return ssh.connect({
+      host: sessionInfo.new_server.networks.v4[0].ip_address,
+      username: 'root',
+      privateKey: sshKey
+    }).then(session => {
+      return session.requestSFTP().then(sftp => {
+        var configReadable = new readable
+        const writeStream = sftp.createWriteStream(configurations[service][config], streamOpts)
+        const stream = new Promise((resolve, reject) => {
+          writeStream.on('close', () => {
+            console.log(config + " transferred succesfully")
+            resolve()
+          })
+        })
+
+        configReadable.push(JSON.stringify(sessionInfo[service][config], null, 2) + '\n')
+        configReadable.push(null)
+        configReadable.pipe(writeStream)
+        return stream
+      }).catch(error => {
+        return console.error(error)
+      })
+    })
+  }
+
+  return readConfig('shadowsocks', 'config.json').then(() => {
+    return writeConfig('shadowsocks', 'config.json')
+  }).then(() => {
+    return serviceRestart('shadowsocks-libev', sessionInfo)
+  }).then(results => {
+    console.log(results)
+    return results
+  }).catch(error => {
+    console.error(error)
+  })
+}
+
 
 //service restart
-function serviceRestart(service) {
-  sequestOpts.command = 'service ' + service + ' restart';
-  sshCommand(newServer.networks.v4[0].ip_address, sequestOpts);
-};
-
-//update domain to reflect new IP.
-function updateDomainRecords(domain) {
-
- client.domainRecordGetAll(domain).then(function(domainRecords) {
-   newRecord = {
-     "type": "A",
-     "data": newServer.networks.v4[0].ip_address,
-     "name": "@"
-   }
-   targetRecord = domainRecords.filter(function(el) {
-     return domainRecords;
-   });
-   if (targetRecord != null){
-     targetRecord = targetRecord.filter(function(el) {
-       return el.type === "A";
-     });
-     if (targetRecord[0].data === oldServer[0].networks.v4[0].ip_address) {
-       console.log("the targetRecord and current server record match, updating record.");
-     } else {
-       throw "DNS Records do not match, skipping DNS record update."
-     }
-     client.domainRecordEdit(domain,targetRecord[0].id,newRecord).then(function (updatedRecord){
-       console.log("The new record is : \n", updatedRecord);
-     })
-   } else {
-     throw "The record details could not be found, skipping DNS record update"
-   }
- });
-}
-// Verify domain record has been successfully updated against DO DNS servers
-
-//helper function to push stream into a string variable
- function streamToString (stream, result, resultKey, callback) {
-   var data = '';
-   var chunk;
-
-   stream.on('readable', function() {
-       while ((chunk=this.read()) != null) {
-           data += chunk;
-           console.log("Reading data . . .");
-       }
-   });
-
-   stream.on('end', function() {
-     data = JSON.parse(data);
-     console.log("reading complete");
-     result[resultKey] = data;
-     callback();
-   });
- };
-
- //genereic sequest callback
- function sequestResponse (error, stdout) {
-    if (error) {
-      throw error;
-    } else {
-      console.log(stdout);
-      return true;
-    }
-  };
-
-  function sshCommand (server, options) {
-    sequest('root@' + server, sequestOpts, sequestResponse);
-  };
-
-
-  function isActive(server, delay, timeout) {
-    client.dropletGet(server).catch(function (error){
-      if (error === 'Error: Request Failed: Error: read ECONNRESET') {
-        console.error('there was an error, trying again:', error);
+function serviceRestart(service, sessionInfo) {
+  console.log('restarting ' + service)
+  return ssh.connect({
+    host: sessionInfo.new_server.networks.v4[0].ip_address,
+    username: 'root',
+    privateKey: sshKey
+  }).then(results => {
+    return ssh.execCommand('service ' + service + ' restart && service ' + service + ' status')
+  }).then(results => {
+    if (results.hasOwnProperty('stdout') || results.hasOwnProperty('stderr')) {
+      if (results.hasOwnProperty('stdout')) {
+        console.log('STDOUT: ' + results.stdout)
+        return sessionInfo
       } else {
-      console.error('there was an error, aborting:', error);
-    }
-    }).then(function statusIs(droplet){
-      if (droplet.status !='active') {
-        console.log(droplet.status);
-        console.log('not yet active');
-        isActive(server);
-      } else {
-        console.log('the server status is ' + droplet.status);
+        return console.error('STDERR: ' + results.stderr)
       }
-    });
-  };
+    }
+  }).catch(error => {
+    return console.error(error)
+  })
+}
 
-*/
+function updateDomainRecords(sessionInfo) {
+  console.log('updating domain record . . .')
+  const domain = sessionInfo.domain
+  return client.domainRecordGetAll(domain).then(domainRecords => {
+    const newRecord = {
+      "type": "A",
+      "data": sessionInfo.new_server.networks.v4[0].ip_address,
+      "name": "@"
+    }
+    var targetRecord = domainRecords.filter(el => {
+      return domainRecords
+    })
+    if (targetRecord != null) {
+      targetRecord = targetRecord.filter(el => {
+        return el.type === "A"
+      })
+      if (targetRecord[0].data === sessionInfo.old_server.networks.v4[0].ip_address) {
+        console.log('the targetRecord and current server record match, updating record')
+      } else {
+        console.error('DNS Records do not match, skipping DNS record update.')
+      }
+      return client.domainRecordEdit(domain, targetRecord[0].id, newRecord).then(updatedRecord => {
+        sessionInfo.domain = updatedRecord
+        console.error('The new record is : \n', sessionInfo.domain)
+      })
+    } else {
+      console.error('The record details could not be found, skipping DNS record update')
+    }
+  }).then(() => sessionInfo).catch(error => console.error(error))
+}
+
+getSessionInfo()
+  .then(results => getCurrentServer(results))
+  .then(results => startNewServer(results))
+  .then(results => checkConnection(results))
+  .then(results => getShadowsocks(results))
+  .then(results => updateDomainRecords(results))
+  .then(results => console.log('redeployment complete: \n' + results.new_server.networks.v4[0].ip_address))
